// src/app/shared/enums/page-name.enum.ts
export enum PageName {
  CONNECTION_PROFILE = 'connection_profile',
  CONFIGURATION_TEMPLATE = 'configuration_template',
  DEVICE = 'device',
  DEVICE_TEMPLATE = 'device_template',
  HARDWARE = 'hardware',
  LOGIN_PROFILES = 'login_profiles',
  ROLES = 'roles',
  IMAGES = 'images',
  ICON = 'icon'
}